requirements.txt
flask
google-genai
python-dotenv
requests
pandas
pandas_ta
